20230424/1/moodclient/pyproject.toml
[project]
name = "MoodClient"
version = "0.0.1"
description = "MOOD Client"
requires-python = ">=3.10"
authors = [{name = "Basalov Iaroslav", email = "s02200273@gse.cs.msu.ru"}]
keywords = ["mood", "project"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
]

dependencies = ["python-cowsay>=1.1.0"]

[project.urls]
repository = "https://github.com/Marilius/pythonprac"

[build-system]
requires = ["alabaster", "Babel", "build", "doit", "flake8", "pybabel", "pycodestyle", "pydocstyle", "pyflakes", "pyreadline3", "python-cowsay", "setuptools", "Sphinx"]
build-backend = "setuptools.build_meta"

[project.scripts]
moodclient = "moodclient.__main__:run"
